refactor: Controller 제거
